Extract CreateChallengePanel from OpenChallengeModal (053 foundational)
Lift the consolidated open-challenge create form (amount hero + pad, memo,
self/third-party/oracle resolution with the stepped market search, deadlines,
submit, result) out of OpenChallengeModal into a reusable CreateChallengePanel
that renders inline (embedded, for the home screen) or inside the modal shell.
OpenChallengeModal is now a thin wrapper. Add a focused panel test (embedded
render, self-resolve submit + onDone, oracle gating + market-search step,
initialResolutionType preselect). Finish the panel-level WIP test debt for the
removed subtitle/caption. Resolves analyze finding A1 (Wagers nav icon → 'ticket').

Marks T003-T007 complete.
