+++ specs/053-home-create-challenge/tasks.md
@@ -43,11 +43,11 @@ Web app — frontend only. All source under `frontend/src/`.
 
 **⚠️ CRITICAL**: The home screen and the Wagers open-challenge entry both depend on the extracted panel.
 
-- [ ] T003 [P] Write failing tests in `frontend/src/test/CreateChallengePanel.test.jsx` per `contracts/create-challenge-panel.md`: renders in `embedded` and modal modes; self / third-party / oracle paths reachable; oracle pill locked off-network and selectable on-network → market-search step → market card + side; submit passes the correct settlement config to the mocked `useOpenChallengeCreate`; success shows the claim-code result and calls `onDone`; zero amount keeps create disabled. Mock `useChainTokens` + `PolymarketBrowser` as in the deleted oracle test.
-- [ ] T004 Extract the consolidated `MakerPanel` from `frontend/src/components/fairwins/OpenChallengeModal.jsx` into `frontend/src/components/fairwins/CreateChallengePanel.jsx` — export `CreateChallengePanel({ embedded=false, onClose, onDone, initialResolutionType })`; move all create state/logic (amount, memo, resolution incl. oracle path + market step, deadlines, submit, result view). Render the `ClaimCodeResultPanel` result inline and call `onDone` on success.
-- [ ] T005 Reduce `frontend/src/components/fairwins/OpenChallengeModal.jsx` to a thin modal shell (backdrop + header + close) wrapping `<CreateChallengePanel embedded={false} … />`, preserving its `isOpen`/`onClose`/`initialResolutionType` props and the header InfoTip.
-- [ ] T006 Make T003 pass — run `npm run test:frontend -- CreateChallengePanel`; confirm both render modes and all three resolution paths work and submit unchanged.
-- [ ] T007 [P] Finish the WIP test debt at the modal/panel level (subtitle + caption removals): update `frontend/src/test/claimCode/OpenChallengeModal.test.jsx` (drop the removed "About: Stake — each side" caption assertion; keep the header "About open challenges" InfoTip), `frontend/src/test/claimCode/publicModeState.test.jsx` (subtitle text removed; InfoTip still present), and `frontend/src/test/claimCode/OpenChallengeCodeBackup.test.jsx` as needed so the Open Challenge modal suite is green.
+- [X] T003 [P] Write failing tests in `frontend/src/test/CreateChallengePanel.test.jsx` per `contracts/create-challenge-panel.md`: renders in `embedded` and modal modes; self / third-party / oracle paths reachable; oracle pill locked off-network and selectable on-network → market-search step → market card + side; submit passes the correct settlement config to the mocked `useOpenChallengeCreate`; success shows the claim-code result and calls `onDone`; zero amount keeps create disabled. Mock `useChainTokens` + `PolymarketBrowser` as in the deleted oracle test.
+- [X] T004 Extract the consolidated `MakerPanel` from `frontend/src/components/fairwins/OpenChallengeModal.jsx` into `frontend/src/components/fairwins/CreateChallengePanel.jsx` — export `CreateChallengePanel({ embedded=false, onClose, onDone, initialResolutionType })`; move all create state/logic (amount, memo, resolution incl. oracle path + market step, deadlines, submit, result view). Render the `ClaimCodeResultPanel` result inline and call `onDone` on success.
+- [X] T005 Reduce `frontend/src/components/fairwins/OpenChallengeModal.jsx` to a thin modal shell (backdrop + header + close) wrapping `<CreateChallengePanel embedded={false} … />`, preserving its `isOpen`/`onClose`/`initialResolutionType` props and the header InfoTip.
+- [X] T006 Make T003 pass — run `npm run test:frontend -- CreateChallengePanel`; confirm both render modes and all three resolution paths work and submit unchanged.
+- [X] T007 [P] Finish the WIP test debt at the modal/panel level (subtitle + caption removals): update `frontend/src/test/claimCode/OpenChallengeModal.test.jsx` (drop the removed "About: Stake — each side" caption assertion; keep the header "About open challenges" InfoTip), `frontend/src/test/claimCode/publicModeState.test.jsx` (subtitle text removed; InfoTip still present), and `frontend/src/test/claimCode/OpenChallengeCodeBackup.test.jsx` as needed so the Open Challenge modal suite is green.
 
 **Checkpoint**: `CreateChallengePanel` is extracted, reused by the modal, and green.
